removed contact forms and show current selected contacts in form vendor
